+++ NEWS.md
@@ -4,6 +4,7 @@
 * Reduced animated tile marker opacity during terrain editing (by Huy Vũ, #4449)
 * Fixed ability to change properties after deselecting the current object (#4440)
 * Fixed Properties view getting stuck with updates disabled (#4506)
+* Fixed copy/paste of list properties losing item types (#4514)
 * Fixed locale-aware parsing of numbers in expression-capable spin boxes
 * Fixed Tile Animation Editor update after tileset image reload (#3923)
 * Fixed runtime language switching in many editor widgets and models (by SIDDHAARTHAA, #4411)